added footer and added about to order page and extra about one
